Multiple bug fixes & code cleanup
Fixed some bugs that were introduced when I changed the way the timer worked.
New method #setEndColor(color) can be used to set the color the embed will end on.
New event preSend is fired in #build() before the first embed is sent in the channel. Has params of reactions, which contains the enabled and disabled reactions for the builder.